parse-data

+++ Controller/User.controller.js
@@ -1,8 +1,12 @@
 import User from "../Model/User.model.js";
 
-const getAllUsers = async (req, res) => {
+export const getAllUsers = async (req, res) => {
     try {
-        const users = await User.find()
+        const users = await User.find({})
+
+        if (!users) {
+            throw new Error("No user signed in yet!")
+        }
         res.json(users);
     }
     catch (err) {
@@ -12,12 +16,16 @@ const getAllUsers = async (req, res) => {
     }
 }
 
-const getUserById = async (req, res) => {
+export const getUserById = async (req, res) => {
+
     try {
-        const user = await User.findById(req.params.id)
+        const user = await User.findOne({ userId: req.params.id })
+
+        if (!user) {
+            throw new Error("User not found!")
+        }
         res.json(user);
     }
-
     catch (err) {
         res.status(500).json({
             error: err.message
@@ -25,8 +33,15 @@ const getUserById = async (req, res) => {
     }
 }
 
-const createUser = async (req, res) => {
+export const createUser = async (req, res) => {
+
     try {
+        const userCheck = await User.findOne({ email: req.body.email })
+
+        if (userCheck) {
+            throw new Error("User already exists!")
+        }
+
         const user = new User(req.body)
         await user.save()
 
@@ -40,12 +55,13 @@ const createUser = async (req, res) => {
     }
 }
 
-const updateUser = async (req, res) => {
+export const updateUser = async (req, res) => {
     try {
-        const user = await User.findByIdAndUpdate(req.params.id, req.body, {
+        const user = await User.findOne({ userId: req.params.id },{
             new: true,
         })
-        res.json(user);
+
+        res.json(user)
     }
 
     catch (err) {
@@ -55,19 +71,14 @@ const updateUser = async (req, res) => {
     }
 }
 
-const deleteUser = async (req, res) => {
+export const deleteUser = async (req, res) => {
     try {
-        await User.findByIdAndDelete(req.params.id)
+        await User.findByIdAndDelete({ userId: req.params.id })
         res.json({ message: "User deleted" });
     }
-
     catch (err) {
         res.status(500).json({
             error: err.message
         })
     }
 }
-
-const userRoutes = [deleteUser, getAllUsers, getUserById, updateUser, createUser]
-
-export default userRoutes;
